IncomeCostsFragment and RecordDialog were implemented.

+++ app/src/main/java/com/example/budgetcontrol/db/dao/TargetDao.kt
@@ -12,11 +12,11 @@ interface TargetDao {
     fun update(target: Target)
 
     @Query("select amount from target where id = :id")
-    fun getTargetAmount(id: Int): Int
+    fun getTargetAmount(id: Int): Float
 
     @Query("select collected_amount from target where id = :id")
-    fun getTargetCollectedAmount(id: Int): Int
+    fun getTargetCollectedAmount(id: Int): Float
 
     @Query("update target set collected_amount = :amount where id = :id")
-    fun updateTargetCollectedAmount(id: Int, amount: Int)
+    fun updateTargetCollectedAmount(id: Int, amount: Float)
 }
