+++ app/src/main/java/com/example/budgetcontrol/db/model/GaussNumber.kt
@@ -7,8 +7,8 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "gauss_numbers")
 data class GaussNumber(
     @PrimaryKey(autoGenerate = true) val id: Int,
-    val value: Int,
+    val value: Float,
     @ColumnInfo(name = "is_collected") val isCollected: Boolean
 ){
-    constructor(value: Int, isCollected: Boolean) : this(0, value, isCollected)
+    constructor(value: Float, isCollected: Boolean) : this(0, value, isCollected)
 }
